feat: VM interpreter refactoring and BCHN script test integration
Refactor the script interpreter to use composite error returns and
align with BCHN's script test battery as a foundation for implementing
pending CHIPs (64-bit integers, Native Introspection, BigInt, VM Limits).

Interpreter refactoring:
- Add op_result struct pairing generic error categories with failing opcode
- Replace ~90 op-specific error codes with 10 generic categories
- Convert all interpreter op_xxx functions to return {category, opcode}
- Add is_signature_result() to centralize fatal vs non-fatal sig errors
- Fix BCH multisig Satoshi bug handling (always discard dummy element)
- Fix P2SH push-only check to return sig_pushonly (matching BCHN)
- Deduplicate script::verify by delegating to free verify() function

New feature implementations:
- Add verify_schnorr() using secp256k1_schnorr_verify
- Implement Schnorr signature path in checksig (64-byte endorsements)
- Implement full checkdatasig/checkdatasigverify with DER and Schnorr support
- Add NULLFAIL enforcement in checkdatasig (bch_daa_cw144)
- Add sig_badlength rejection for 65-byte endorsements in checkmultisig
- Add MINIMALNUM validation in CLTV/CSV opcodes (bch_mersenne)
- Fix CLTV MINIMALNUM check ordering (before is_final)
- Fix checkmultisig error propagation for pop_int32 failures

Native introspection opcodes:
- Implement all 14 introspection opcodes (OP_INPUTINDEX through OP_UTXOBYTECODE)
- Add CashTokens token prefix parsing for output_token_* opcodes
- Add script_execution_context for pre-computed transaction introspection data
- Propagate context across program constructors

Signature validation (BCHN-compatible):
- Add is_valid_signature_encoding(): pure DER format check matching BCHN
- Add check_low_s(): secp256k1-based LOW_S validation
- Add check_transaction_signature_encoding(): DER + LOW_S + hashtype + forkid
- Add check_data_signature_encoding(): for OP_CHECKDATASIG (no sighash byte)
- Schnorr 65-byte endorsements skip DER/LOW_S checks
- SIGPUSHONLY check before script evaluation in verify()
- MINIMALIF enforcement in op_if/op_notif

rule_fork refactoring:
- Each BCH consensus feature gets its own bit in forks_t (uint64_t)
- Upgrade groupings as constexpr OR combinations
- Add forks_t type alias and are_enabled() variadic template helper
- Consolidated missing_forkid into must_use_forkid (BCHN alias)
- Regenerated C-API error codes

Program and verify:
- program constructor accepts value param for SIGHASH_FORKID amount
- verify() passes value through to program for correct sighash computation
- Propagate metrics across program constructors (input -> prevout -> P2SH)

INPUT_SIGCHECKS (Fermat upgrade):
- Implement per-input sigchecks density limit: scriptSig.size() >= sigChecks * 43 - 60
- ECDSA CHECKMULTISIG counts nKeysCount (not 1) when any sig is non-null
- Schnorr CHECKMULTISIG counts 1 per verified signature
- Add sigchecks_input_density_factor/offset constants

BigInt (big_number) implementation:
- Add big_number class wrapping GMP (mpz_t) for arbitrary-precision integers
- Support construction from decimal strings, hex strings, and script number serialization
- Implement arithmetic (+, -, *, /, %), bitwise (&, |, ^, <<, >>), and comparison operators
- Add math_modulo (floored), pow, pow_mod operations
- Add from_hex() for OpenSSL-format hex parsing
- Add 23K+ auto-generated tests from BCHN test vectors (bigint_test_vectors.json)
- Add 2.5K+ OpenSSL-format tests (sum, mul/square/quotient, exp, mod, shift)
- Test generators use static arrays + for loops to minimize linker symbols

VM Limits tests:
- Auto-generate 165 VM limits tests from BCHN (128 success + 15 fail + 24 checker)
- Auto-resolve CHECKSIG tests with real transaction context and UTXO cache
- Support CashTokens token prefix parsing in test scriptPubKey

VM Limits metrics fixes:
- Move sig_checks/hash_iterations tallying into op_check_sig_common and
  op_check_data_common (before NULLFAIL check), matching BCHN behavior
- Fix CHECKDATASIG: tally sig_checks(1) and hash_iterations for any
  non-empty signature, regardless of verification success/failure
- Fix CHECKSIG: same tallying fix for non-empty failed signatures
- Callers (op_check_sig, op_check_data_sig, etc.) now only tally add_op_cost
- Fix version_ not propagated in program copy/move constructors (non-BCH)
- Replace DummySigChecker with StandaloneSigChecker in eval_script_with_metrics
  so CHECKDATASIG does real secp256k1 verification in cross-check tests
- Add CHECKDATASIG failed-verification cross-check test (native vs BCHN)
- Add CHECKDATASIG metrics test for non-empty failed sig (sig_checks=1)
  and empty sig (sig_checks=0)

Additional fixes:
- Fix op_if/op_notif: return unbalanced_conditional on empty stack
- Fix op_check_sequence_verify: return insufficient_main_stack on empty stack
- Fix op_check_sig_verify: use is_signature_result pattern
- Fix parse_endorsement: add const& overload to prevent move in ECDSA multisig retry loop

Test infrastructure:
- Integrate BCHN script_tests.json with generate_script_tests.py
- Generate individual Catch2 TEST_CASEs split across 18 files (100/file)
- Test names include expected result: [Should Pass] / [Should Fail: ERROR]
- Generator validates overrides don't change OK<->FAIL semantics
- Generator handles [[amount], ...] format from BCHN for SIGHASH_FORKID tests
- Add override annotation comments to generated test files
- bchn_script_test struct supports amount field for FORKID tests
- Update SIG_DER mapping to invalid_signature_lax_encoding
- Enable 29 previously-skipped tests (CLEANSTACK, 64-bit arithmetic, expand_forks)
- Add OVERFLOW error mapping to generator, 4 error overrides for 64-bit overflow
- All 1670 BCHN script tests passing (149 overrides, 76 skipped)

+++ src/c-api/include/kth/capi/chain/transaction.h
@@ -105,13 +105,13 @@ kth_input_list_t kth_chain_transaction_inputs(kth_transaction_t transaction);
 KTH_EXPORT
 uint8_t const* kth_chain_transaction_to_data(kth_transaction_t transaction, kth_bool_t wire, kth_size_t* out_size);
 
-// code verify(transaction const& tx, uint32_t input_index, uint32_t forks, script const& input_script, script const& prevout_script, uint64_t /*value*/);
+// code verify(transaction const& tx, uint32_t input_index, uint64_t forks, script const& input_script, script const& prevout_script, uint64_t /*value*/);
 KTH_EXPORT
-kth_error_code_t kth_chain_transaction_verify(kth_transaction_t transaction, uint32_t input_index, uint32_t forks, kth_script_t input_script, kth_script_t prevout_script, uint64_t value);
+kth_error_code_t kth_chain_transaction_verify(kth_transaction_t transaction, uint32_t input_index, uint64_t forks, kth_script_t input_script, kth_script_t prevout_script, uint64_t value);
 
-// code verify(transaction const& tx, uint32_t input, uint32_t forks);
+// code verify(transaction const& tx, uint32_t input, uint64_t forks);
 KTH_EXPORT
-kth_error_code_t kth_chain_transaction_verify_transaction(kth_transaction_t transaction, uint32_t input, uint32_t forks);
+kth_error_code_t kth_chain_transaction_verify_transaction(kth_transaction_t transaction, uint32_t input, uint64_t forks);
 
 KTH_EXPORT
 kth_error_code_t kth_chain_transaction_create_template_with_change_ratios(
